move loose api utils to utils/api
