+++ pages/api/attest/github.js
@@ -3,7 +3,10 @@ import withMethods from '../../../util/api/withMethods'
 import withAuth from '../../../util/api/withAuth'
 import withValidParams from '../../../util/api/withValidParams'
 import { githubClientId } from '../../../util/constants'
-import { findMemberByAddress, updateMemberGithub } from '../../../util/fauna'
+import {
+  findMemberByAddress,
+  updateMemberGithub,
+} from '../../../util/api/fauna'
 
 async function getAccessToken(code) {
   const resp = await fetch('https://github.com/login/oauth/access_token', {
