+++ pages/api/rollup.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-await-in-loop */
 import withAuth from '../../util/api/withAuth'
 import withMethods from '../../util/api/withMethods'
-import { findCommunityByAddress } from '../../util/fauna'
+import { findCommunityByAddress } from '../../util/api/fauna'
 
 // TODO: break out into a set of util functions:
 const ROLLUP_FUNCTIONS = {
