+++ pages/api/save-integration-config.js
@@ -4,7 +4,7 @@ import withValidParams from '../../util/api/withValidParams'
 import {
   findCommunityByAddress,
   updateCommunitySnapshotEns,
-} from '../../util/fauna'
+} from '../../util/api/fauna'
 
 const handler = async (req, res, { auth: { address } }) => {
   // Get config:
